+++ linux_parsers/parsers/filesystem/__init__.py
@@ -1,5 +1,5 @@
 from linux_parsers.parsers.filesystem.df import parse_df
-from linux_parsers.parsers.filesystem.dpkg import parse_dpkg_l
+from linux_parsers.parsers.filesystem.packages import parse_dpkg_l, parse_rpm_qa, parse_yum_list_installed
 from linux_parsers.parsers.filesystem.du import parse_du
 from linux_parsers.parsers.filesystem.fdisk import parse_fdisk
 from linux_parsers.parsers.filesystem.ls import parse_ls
@@ -10,6 +10,8 @@ __all__ = [
     "parse_df",
     "parse_ls",
     "parse_dpkg_l",
+    "parse_rpm_qa",
+    "parse_yum_list_installed",
     "parse_fdisk",
     "parse_stat",
     "parse_du",