+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.2.7 <date>
-- Added command parser: `rpm -qa`
+- Added command parser: `rpm -qa`.
+- Added command parser: `yum list installed`
+- Changed `dpkg.py` to `packages.py` to have all packages parsers in one place.
 
 ### 0.2.6 08.03.2025